requirements.txt
boto3
pyfakefs
pyhcl
pylint
yara-python==3.6.3
python-dotenv
